ai-gateway: realtime route as thin handler + service + transport
